refactor(temp-mail): updating some function and constant names
Some constant and functions names has been updated and some functions has been removed or modified to better understanding purposes. Besides, modifying the functions, we can improve the behavior of the temp mail.
