+++ src/components/temp-email/hooks/useTempMailReducer.ts
@@ -1,12 +1,11 @@
 import { useCallback, useReducer } from 'react';
-import { ActionType, ActionTypes, MessageObjProps } from '../types/types';
+import { ActionType, ActionTypes, MessageObjProps, UserProps } from '../types/types';
 
 const reducer = (state, action: ActionType) => {
   switch (action.type) {
-    case ActionTypes.SET_EMAIL:
-      return { ...state, email: action.payload };
-    case ActionTypes.SET_TOKEN:
-      return { ...state, token: action.payload };
+    case ActionTypes.SET_USER:
+      return { ...state, user: action.payload };
+
     case ActionTypes.SET_BORDER_COLOR:
       return { ...state, borderColor: action.payload };
     case ActionTypes.SET_OPENED_MESSAGES:
@@ -27,8 +26,7 @@ const reducer = (state, action: ActionType) => {
 };
 
 const initialState = {
-  email: undefined,
-  token: undefined,
+  user: undefined,
   borderColor: false,
   openedMessages: 0,
   isRefreshed: false,
@@ -41,12 +39,8 @@ const initialState = {
 export const useTempMailReducer = () => {
   const [state, dispatch] = useReducer(reducer, initialState);
 
-  const setEmail = useCallback((email: string | undefined) => {
-    dispatch({ type: 'SET_EMAIL', payload: email });
-  }, []);
-
-  const setToken = useCallback((token: string | undefined) => {
-    dispatch({ type: 'SET_TOKEN', payload: token });
+  const setUser = useCallback((user: UserProps | undefined) => {
+    dispatch({ type: 'SET_USER', payload: user });
   }, []);
 
   const setBorderColor = useCallback((borderColor: boolean) => {
@@ -79,8 +73,7 @@ export const useTempMailReducer = () => {
 
   return {
     state,
-    setEmail,
-    setToken,
+    setUser,
     setBorderColor,
     setOpenedMessages,
     setIsRefreshed,
