+++ src/components/temp-email/services/temp-mail.service.ts
@@ -1,17 +1,19 @@
 import axios from 'axios';
 import { MessageObjProps } from '../types/types';
 
-export const EMAIL_STORAGE_KEY = 'email';
+export const EMAIL_STORAGE_KEY = 'temp-mail-user-data';
 export const SETUP_TIME_STORAGE_KEY = 'setupTime';
 export const INBOX_STORAGE_KEY = 'inbox';
 
 export const TIME_NOW = new Date().getTime();
 export const MAX_HOURS_BEFORE_EXPIRE_EMAIL = 5 * 60 * 60 * 1000;
 
-const getEmail = async () => {
+const fetchNewEmail = async () => {
   const email = await axios.get(`${window.origin}/api/temp-mail/create-email`);
 
-  return email.data;
+  const { data: userInfo } = email;
+
+  return userInfo;
 };
 
 const fetchInbox = async (email: string, token: string) => {
@@ -30,12 +32,6 @@ const getMessageData = async (email: string, token: string, messageId: string) =
   return messageData;
 };
 
-const createEmail = async () => {
-  const fetchEmail = await getEmail();
-
-  return fetchEmail;
-};
-
 const fetchAndFormatInbox = async (email: string, userToken: string): Promise<MessageObjProps[] | undefined> => {
   try {
     const inbox = await fetchInbox(email, userToken);
@@ -46,19 +42,7 @@ const fetchAndFormatInbox = async (email: string, userToken: string): Promise<Me
       throw new Error('Auto fetching email');
     }
 
-    if (inbox.length > 0) {
-      const messages = inbox.map((item, index) => {
-        console.log({ seen: item.seen });
-        return {
-          ...item,
-          opened: item.seen,
-        };
-      });
-
-      return messages;
-    }
-
-    return [];
+    return inbox;
   } catch (err) {
     const error = err as Error;
     throw new Error(error.message);
@@ -72,4 +56,4 @@ function removeLocalStorage() {
   localStorage.removeItem('selectedMessage');
 }
 
-export { createEmail, fetchInbox, getMessageData, fetchAndFormatInbox, removeLocalStorage };
+export { fetchNewEmail, fetchInbox, getMessageData, fetchAndFormatInbox, removeLocalStorage };
